standalone verify function
The 'verify' class method of KeyPair is more typically used in the
context where the signing key is not known.
For this reason, the 'verify' method is now a standalone function which
takes an verifying (public) nkey, and the class method (of the same
name) calls the standalone function.
Additionally, since it is no longer guaranteed thatt the nkey will be
well-formed, validation is performed.